vision-service/docker/yolo/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
Pillow>=10.3.0
